--- imports/plugins/core/permissions/register.js
@@ -1,8 +1,0 @@
-import { Reaction } from "/server/api";
-
-Reaction.registerPackage({
-  label: "Reaction",
-  name: "reaction-permissions",
-  icon: "fa fa-lock",
-  autoEnable: true
-});
