+++ imports/plugins/core/ui/client/components/permission/index.js
@@ -1,0 +1,1 @@
+export * from "./permission";
